+++ config/config.exs
@@ -1,3 +1,5 @@
 import Config
 
 import_config "#{Mix.env()}.exs"
+
+config :tesla, :adapter, Tesla.Adapter.Hackney